pcb_model/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pillow
python-multipart
ultralytics
httpx
pydantic
inference-sdk
